fanotify03: Test FAN_DENY with custom errno
Add a test case for returning a custom errno from FAN_PRE_ACCESS
and FAN_OPEN_EXEC_PERM events.

The test is changed to use class FAN_CLASS_PRE_CONTENT, which is
required for responsing with FAN_DENY_ERRNO.

Instead of testing for kernel suppport for FAN_DENY_ERRNO, we test for
kernel support for FAN_PRE_ACCESS events, which is easier.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>