fanotify01: Test setting two marks on different filesystems
When tested fs has zero fsid (e.g. fuse) and events are reported
with fsid+fid, watching different filesystems with the same group is
not supported.

Try to setup a bogus mark on test tmp dir, in addition to the mark
on the tested filesystem, to check if marks on different filesystems
are supported.

Run on all filesystem to test both fs with zero fsid (e.g. fuse) and
fs with valid fsid.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>