fanotify13: Verify that we did not get an extra event
For example, verify that we did not get an event on a directory object
without requesting FAN_ONDIR.  Also, report a test failure if no events
received at all instead of blocking on read of fanotify_fd.

With this change, the test fails with overlayfs variants over btrfs,
because the size of fid of overalyfs over btrfs is about 90 bytes and
the events on the three objects do not all fit into a single 256 bytes
buffer read. Increase the size of the events buffer to fix this failure.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>